requirements.txt
torch
nltk
gensim
matplotlib
wordcloud
tensorboard
jupyter
hydra-core
omegaconf
scikit-learn
